Alignment of last paragraph
Could you align the last paragraph with the Resident Guidance Counselors paragraph. I'm not sure if I corrected this right. Could you check? Also, could you add a link to the Testimonials page on the following words: read what some of the parents and students have to say. Thanks Shubhie